MSC 5.1/clang warning/error round
no new warnings with MSC 5.1 /W3 and no "make clean && make verify" change to previous version

+++ src/start2.c
@@ -349,7 +349,7 @@ void pilotToGameData(uint8 *pilotData)
 }
 
 // 22ec
-int pilotNameInput(int *page, int a, int b, int c, struct Pilot *pilot) {
+int pilotNameInput(int16 *page, int a, int b, int c, struct Pilot *pilot) {
     int blinkToggle;
     int x, y;
     int nameLen;
